requirements.txt
google-genai
python-dotenv
pydantic
pymupdf